Guangzhou Metropolitan Social Space Quality Score Data

Data of Social Space Quality Score in Guangzhou Metropolitan Areas of China (2010). The actual representation of each column is as follows: SName_EN street name; DName_EN district name; SSQ_Score total score of social space quality; PS_Score population stability score; EL_Score educational level score; OH_Score occupational hierarchy score; IL_Score income level score; FPOP_Pro proportion of foreign population; TenantsPro proportion of tenants; NoSchPro proportion of no schooling; PSchPRO proportion of primary school education; JHSchPro proportion of junior high school education; HSchDipPro proportion of high school diploma; CDegreePro proportion of college degree; UnderG_Pro proportion of undergraduate; PostG_Pro proportion of postgraduate; RPSOPMOPro proportion of responsible persons of state organs, party and mass organizations and institutions; PCE_Pro proportion of person in charge of enterprise; ProTechPro proportion of professional and technical personnel; ClerkPro proportion of clerk; BusSer_Pro proportion of business and service personnel; AFAFP_Pro proportion of agriculture, forestry, animal husbandry and fishery personnel; OPTE_Pro proportion of operators of production and transportation equipment; UnemPeoPro proportion of households with unemployed people; B100_Pro proportion of households with below 100 yuan; 100_200Pro proportion of households with 100-200yuan; 200_500Pro proportion of households with 200-500yuan; 500_1000P proportion of households with 500-1000yuan; 1000_1500P proportion of households with 1000-1500yuan; 1500_2000P proportion of households with 1500-2000yuan; 2000_3000P proportion of households with 2000-3000yuan; A3000_Pro proportion of households with above 3000yuan; The subdistrict boundary is drawn by with reference to the Atlas of Community Network ResponsibilityDistrict of Urban Management Division in Guangzhou.

Usage

gzma

Format

gsma: An sf tibble of social space quality score in guangzhou metropolitan areas(2010) with 118 rows and 32 variables, where the last column is geometry.

References

WANG Yang, ZHANG Hong’ou, YE Yuyao, WU Qitao, JIN Lixia. Comprehensive Evaluation and Distribution Pattern of Social Space Quality in Guangzhou, China. Tropical Geography.